+++ CHANGELOG.md
@@ -34,6 +34,8 @@
 
 - Migrate to headless Chrome (see #96).
 
+- Improve performance of Karma (see #97).
+
 - Update dev dependencies to latest version.
 
 ### [v0.2.1](https://github.com/RobotlegsJS/RobotlegsJS/releases/tag/0.2.1) - 2018-09-12